Migrate xml.py into a package
