+++ src/rstxml2db/xml/util.py
@@ -1,0 +1,30 @@
+#
+# Copyright (c) 2016 SUSE Linux GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of version 3 of the GNU General Public License as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, contact SUSE LLC.
+#
+# To contact SUSE about this file by physical or electronic mail,
+# you may find current contact information at www.suse.com
+
+from lxml import etree
+
+
+def quoteparams(args):
+    """Quote parameters with :func:`etree.XSLT.strparam`
+
+    :param args: result from :class:`argparse` parser
+    :return: parameter list with quoted values
+    :rtype: list
+    """
+    return [(p[0], p[1] if p[0].startswith('_') else etree.XSLT.strparam(p[1]))
+            for p in args.params]
